+++ packages/docusaurus-theme-openapi/src/theme/ApiDemoPanel/MethodEndpoint/index.tsx
@@ -7,16 +7,22 @@
 
 import React from "react";
 
+const methodStyle = {
+  borderRadius: "var(--ifm-global-radius)",
+};
+
 function colorForMethod(method: string) {
   switch (method.toLowerCase()) {
     case "get":
-      return "var(--openapi-code-blue)";
+      return "primary";
     case "put":
-      return "var(--openapi-code-orange)";
+      return "warning";
+    case "patch":
+      return "warning";
     case "post":
-      return "var(--openapi-code-green)";
+      return "success";
     case "delete":
-      return "var(--openapi-code-red)";
+      return "danger";
     default:
       return undefined;
   }
@@ -35,7 +41,10 @@ function MethodEndpoint({ method, path }: Props) {
         borderRadius: "var(--openapi-card-border-radius)",
       }}
     >
-      <span style={{ color: colorForMethod(method) }}>
+      <span
+        style={methodStyle}
+        className={"badge badge--" + colorForMethod(method)}
+      >
         {method.toUpperCase()}
       </span>{" "}
       <span>{path.replace(/{([a-z0-9-_]+)}/gi, ":$1")}</span>